优化账户模块。

+++ ranch-account/src/main/java/org/lpw/ranch/account/AccountModel.java
@@ -23,16 +23,16 @@ public class AccountModel extends ModelSupport {
     private String user; // 用户
     private String owner; // 所有者
     private int type; // 类型
-    private int balance; // 余额
-    private int deposit; // 存入总额
-    private int withdraw; // 取出总额
-    private int reward; // 奖励总额
-    private int profit; // 盈利总额
-    private int consume; // 消费总额
-    private int remitIn; // 汇入总额
-    private int remitOut; // 汇出总额
-    private int refund; // 退款总额
-    private int pending; // 待结算总额
+    private long balance; // 余额
+    private long deposit; // 存入总额
+    private long withdraw; // 取出总额
+    private long reward; // 奖励总额
+    private long profit; // 盈利总额
+    private long consume; // 消费总额
+    private long remitIn; // 汇入总额
+    private long remitOut; // 汇出总额
+    private long refund; // 退款总额
+    private long pending; // 待结算总额
     private String checksum; // 校验值
     private String lockId; // 全局锁ID
 
@@ -68,101 +68,101 @@ public class AccountModel extends ModelSupport {
 
     @Jsonable
     @Column(name = "c_balance")
-    public int getBalance() {
+    public long getBalance() {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(long balance) {
         this.balance = balance;
     }
 
     @Jsonable
     @Column(name = "c_deposit")
-    public int getDeposit() {
+    public long getDeposit() {
         return deposit;
     }
 
-    public void setDeposit(int deposit) {
+    public void setDeposit(long deposit) {
         this.deposit = deposit;
     }
 
     @Jsonable
     @Column(name = "c_withdraw")
-    public int getWithdraw() {
+    public long getWithdraw() {
         return withdraw;
     }
 
-    public void setWithdraw(int withdraw) {
+    public void setWithdraw(long withdraw) {
         this.withdraw = withdraw;
     }
 
     @Jsonable
     @Column(name = "c_reward")
-    public int getReward() {
+    public long getReward() {
         return reward;
     }
 
-    public void setReward(int reward) {
+    public void setReward(long reward) {
         this.reward = reward;
     }
 
     @Jsonable
     @Column(name = "c_profit")
-    public int getProfit() {
+    public long getProfit() {
         return profit;
     }
 
-    public void setProfit(int profit) {
+    public void setProfit(long profit) {
         this.profit = profit;
     }
 
     @Jsonable
     @Column(name = "c_consume")
-    public int getConsume() {
+    public long getConsume() {
         return consume;
     }
 
-    public void setConsume(int consume) {
+    public void setConsume(long consume) {
         this.consume = consume;
     }
 
     @Jsonable
     @Column(name = "c_remit_in")
-    public int getRemitIn() {
+    public long getRemitIn() {
         return remitIn;
     }
 
-    public void setRemitIn(int remitIn) {
+    public void setRemitIn(long remitIn) {
         this.remitIn = remitIn;
     }
 
     @Jsonable
     @Column(name = "c_remit_out")
-    public int getRemitOut() {
+    public long getRemitOut() {
         return remitOut;
     }
 
-    public void setRemitOut(int remitOut) {
+    public void setRemitOut(long remitOut) {
         this.remitOut = remitOut;
     }
 
     @Jsonable
     @Column(name = "c_refund")
-    public int getRefund() {
+    public long getRefund() {
         return refund;
     }
 
-    public void setRefund(int refund) {
+    public void setRefund(long refund) {
         this.refund = refund;
     }
 
     @Jsonable
     @Column(name = "c_pending")
-    public int getPending() {
+    public long getPending() {
         return pending;
     }
 
-    public void setPending(int pending) {
+    public void setPending(long pending) {
         this.pending = pending;
     }
